Move unit test to feature test namespace
