+++ packages/framework/tests/Feature/HydeServiceProviderTest.php
@@ -2,17 +2,8 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Testing\Unit;
+namespace Hyde\Framework\Testing\Feature;
 
-use function app;
-use function array_filter;
-use function array_map;
-use function array_values;
-use function basename;
-use function config;
-use function get_class;
-use function get_declared_classes;
-use function glob;
 use Hyde\Facades\Site;
 use Hyde\Framework\Features\DataCollections\DataCollectionServiceProvider;
 use Hyde\Framework\HydeServiceProvider;
@@ -25,6 +16,15 @@ use Hyde\Pages\MarkdownPage;
 use Hyde\Pages\MarkdownPost;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Facades\Artisan;
+use function app;
+use function array_filter;
+use function array_map;
+use function array_values;
+use function basename;
+use function config;
+use function get_class;
+use function get_declared_classes;
+use function glob;
 use function method_exists;
 use function str_starts_with;
