+++ packages/framework/tests/Feature/YamlConfigurationServiceTest.php
@@ -11,7 +11,7 @@ use Illuminate\Support\Facades\Config;
 /**
  * @covers \Hyde\Framework\Services\YamlConfigurationService
  *
- * @see \Hyde\Framework\Testing\Unit\HydeServiceProviderTest as it determines if this service should be booted.
+ * @see \Hyde\Framework\Testing\Feature\HydeServiceProviderTest as it determines if this service should be booted.
  */
 class YamlConfigurationServiceTest extends TestCase
 {
